fix: cross platform support for --cache=fs
